refactor: set up default branch id
