+++ internal/pkg/project/cachefile/doc.go
@@ -1,0 +1,2 @@
+// Package cachefile manages .keboola/project.json file with cached information about project for local commands, without API access.
+package cachefile
