+++ internal/pkg/service/cli/dependencies/cmdlocal.go
@@ -9,11 +9,11 @@ import (
 	"github.com/keboola/keboola-as-code/internal/pkg/filesystem"
 	"github.com/keboola/keboola-as-code/internal/pkg/model"
 	projectPkg "github.com/keboola/keboola-as-code/internal/pkg/project"
+	"github.com/keboola/keboola-as-code/internal/pkg/project/cachefile"
 	projectManifest "github.com/keboola/keboola-as-code/internal/pkg/project/manifest"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/common/configmap"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/common/dependencies"
 	"github.com/keboola/keboola-as-code/internal/pkg/template"
-	"github.com/keboola/keboola-as-code/internal/pkg/template/project"
 	"github.com/keboola/keboola-as-code/internal/pkg/template/repository"
 	"github.com/keboola/keboola-as-code/internal/pkg/utils/errors"
 	"github.com/keboola/keboola-as-code/internal/pkg/version"
@@ -67,7 +67,7 @@ func newLocalCommandScope(ctx context.Context, baseScp BaseScope, hostByFlag con
 		return nil, err
 	}
 
-	fileContent, err := project.Load(ctx, baseScp.Fs())
+	fileContent, err := cachefile.Load(ctx, baseScp.Fs())
 	if err != nil {
 		return nil, err
 	}
